+++ packages/perseus-editor/src/widgets/interactive-graph-editor/interactive-graph-editor.tsx
@@ -405,6 +405,84 @@ class InteractiveGraphEditor extends React.Component<Props> {
                         />
                     </LabeledRow>
                 )}
+                {this.props.correct?.type === "angle" && (
+                    <>
+                        <View style={styles.row}>
+                            <Checkbox
+                                label={
+                                    <LabelSmall>Show angle measures</LabelSmall>
+                                }
+                                checked={
+                                    // Don't show indeterminate checkbox state
+                                    !!this.props.correct?.showAngles
+                                }
+                                onChange={(newValue) => {
+                                    if (this.props.graph?.type === "angle") {
+                                        invariant(
+                                            this.props.correct.type === "angle",
+                                            `Expected graph type to be angle, but got ${this.props.correct.type}`,
+                                        );
+                                        this.props.onChange({
+                                            correct: {
+                                                ...this.props.correct,
+                                                showAngles: newValue,
+                                            },
+                                            graph: {
+                                                ...this.props.graph,
+                                                showAngles: newValue,
+                                            },
+                                        });
+                                    }
+                                }}
+                            />
+                            <InfoTip>
+                                <p>Displays the interior angle measures.</p>
+                            </InfoTip>
+                        </View>
+                        <View style={styles.row}>
+                            <Checkbox
+                                label={
+                                    <LabelSmall>Allow reflex angles</LabelSmall>
+                                }
+                                checked={
+                                    // Don't show indeterminate checkbox state
+                                    !!this.props.correct?.allowReflexAngles
+                                }
+                                onChange={(newValue) => {
+                                    invariant(
+                                        this.props.correct.type === "angle",
+                                        `Expected graph type to be angle, but got ${this.props.correct.type}`,
+                                    );
+                                    invariant(
+                                        this.props.graph?.type === "angle",
+                                        `Expected graph type to be angle, but got ${this.props.graph?.type}`,
+                                    );
+
+                                    const update = {
+                                        allowReflexAngles: newValue,
+                                    };
+
+                                    this.props.onChange({
+                                        correct: {
+                                            ...this.props.correct,
+                                            ...update,
+                                        },
+                                        graph: {
+                                            ...this.props.graph,
+                                            ...update,
+                                        },
+                                    });
+                                }}
+                            />
+                            <InfoTip>
+                                <p>
+                                    Allow students to be able to create reflex
+                                    angles.
+                                </p>
+                            </InfoTip>
+                        </View>
+                    </>
+                )}
                 {this.props.correct?.type === "polygon" && (
                     <>
                         <LabeledRow label="Number of sides:">