+++ packages/perseus-editor/src/widgets/interactive-graph-editor/start-coords/util.ts
@@ -9,6 +9,7 @@ import {
     getQuadraticCoords,
     getSegmentCoords,
     getSinusoidCoords,
+    getClockwiseAngle,
 } from "@khanacademy/perseus";
 import {UnreachableCaseError} from "@khanacademy/wonder-stuff-core";
 
@@ -146,19 +147,17 @@ export const getQuadraticEquation = (startCoords: [Coord, Coord, Coord]) => {
     );
 };
 
-const findAngle = (point1: Coord, point2: Coord) => {
-    const x = point1[0] - point2[0];
-    const y = point1[1] - point2[1];
+export const getAngleEquation = (
+    startCoords: [Coord, Coord, Coord],
+    allowReflexAngles: boolean = false,
+) => {
+    const vertex = startCoords[1];
 
-    return (180 + (Math.atan2(-y, -x) * 180) / Math.PI + 360) % 360;
-};
-
-export const getAngleEquation = (startCoords: [Coord, Coord, Coord]) => {
-    const [point1, vertex, point2] = startCoords;
-
-    const angle = findAngle(point2, vertex) - findAngle(point1, vertex);
-    const roundedAngle = angle.toFixed(0);
-
+    // Get the angle in degrees and round it to the nearest whole number
+    const roundedAngle = getClockwiseAngle(
+        startCoords,
+        allowReflexAngles,
+    ).toFixed(0);
     return `${roundedAngle}\u00B0 angle at (${vertex[0]}, ${vertex[1]})`;
 };