+++ packages/perseus/src/validation.types.ts
@@ -1,7 +1,5 @@
 import type {Coord} from "./interactive2/types";
 import type {
-    PerseusCategorizerWidgetOptions,
-    PerseusCSProgramWidgetOptions,
     PerseusDefinitionWidgetOptions,
     PerseusDropdownWidgetOptions,
     PerseusExplanationWidgetOptions,
@@ -34,13 +32,20 @@ import type {Relationship} from "./widgets/number-line/number-line";
 
 export type UserInputStatus = "correct" | "incorrect" | "incomplete";
 
-export type PerseusCategorizerRubric = PerseusCategorizerWidgetOptions;
-
-export type PerseusCategorizerUserInput = {
+export type PerseusCategorizerRubric = {
+    // The correct answers where index relates to the items and value relates
+    // to the category.  e.g. [0, 1, 0, 1, 2]
     values: ReadonlyArray<number>;
 };
 
-export type PerseusCSProgramRubric = PerseusCSProgramWidgetOptions;
+export type PerseusCategorizerUserInput = {
+    values: PerseusCategorizerRubric["values"];
+};
+// TODO(LEMS-2440): Can possibly be removed during 2440?
+// This is not used for grading at all. The only place it is used is to define
+// Props type in cs-program.tsx, but RenderProps already contains WidgetOptions
+// and is already included in the Props type.
+export type PerseusCSProgramRubric = Empty;
 
 export type PerseusCSProgramUserInput = {
     status: UserInputStatus;