+++ packages/perseus/src/widgets/interactive-graph.tsx
@@ -31,6 +31,7 @@ import {getInteractiveBoxFromSizeClass} from "../util/sizing-utils";
 
 import {StatefulMafsGraph} from "./interactive-graphs";
 import interactiveGraphValidator from "./interactive-graphs/interactive-graph-validator";
+import {getClockwiseAngle} from "./interactive-graphs/math";
 
 import type {StatefulMafsGraphType} from "./interactive-graphs/stateful-mafs-graph";
 import type {QuadraticGraphState} from "./interactive-graphs/types";
@@ -43,7 +44,7 @@ import type {
     PerseusGraphTypeSegment,
     PerseusInteractiveGraphWidgetOptions,
 } from "../perseus-types";
-import type {ChangeHandler, Widget, WidgetExports, WidgetProps} from "../types";
+import type {ChangeHandler, WidgetExports, WidgetProps} from "../types";
 import type {
     QuadraticCoefficient,
     Range,
@@ -1206,7 +1207,7 @@ class LegacyInteractiveGraph extends React.Component<Props, State> {
                 });
 
                 const getAngle = function (a: number, vertex, b: number) {
-                    const angle = GraphUtils.findAngle(
+                    const angle = GraphUtils.findAngleDeprecated(
                         coords[rel(a)],
                         coords[rel(b)],
                         coords[rel(vertex)],
@@ -1243,7 +1244,7 @@ class LegacyInteractiveGraph extends React.Component<Props, State> {
                         Math.sin((innerAngles[2] * Math.PI) / 180)) *
                     knownSide;
 
-                const outerAngle = GraphUtils.findAngle(
+                const outerAngle = GraphUtils.findAngleDeprecated(
                     coords[rel(1)],
                     coords[rel(-1)],
                 );
@@ -1287,7 +1288,7 @@ class LegacyInteractiveGraph extends React.Component<Props, State> {
                 // Solve for angle by using the law of cosines
                 const innerAngle = lawOfCosines(sides[0], sides[2], sides[1]);
 
-                const outerAngle = GraphUtils.findAngle(
+                const outerAngle = GraphUtils.findAngleDeprecated(
                     coords[rel(1)],
                     coords[rel(-1)],
                 );
@@ -1796,7 +1797,7 @@ class LegacyInteractiveGraph extends React.Component<Props, State> {
     }
 }
 
-class InteractiveGraph extends React.Component<Props, State> implements Widget {
+class InteractiveGraph extends React.Component<Props, State> {
     legacyGraphRef = React.createRef<LegacyInteractiveGraph>();
     mafsRef = React.createRef<StatefulMafsGraphType>();
 
@@ -2338,7 +2339,8 @@ class InteractiveGraph extends React.Component<Props, State> implements Widget {
             throw makeInvalidTypeError("getAngleEquationString", "angle");
         }
         const coords = InteractiveGraph.getAngleCoords(props.graph, props);
-        const angle = GraphUtils.findAngle(coords[2], coords[0], coords[1]);
+        const allowReflexAngles = props.graph.allowReflexAngles;
+        const angle = getClockwiseAngle(coords, allowReflexAngles);
         return (
             angle.toFixed(0) +
             "\u00B0 angle" +