+++ packages/perseus/src/widgets/interactive-graphs/graphs/point.tsx
@@ -46,6 +46,15 @@ export function UnlimitedPointGraph(props: PointGraphProps) {
         height,
     ]);
     const [[left, top]] = useTransformVectorsToPixels([minX, maxY]);
+    const itemsRef = React.useRef<Array<SVGElement | null>>([]);
+
+    React.useEffect(() => {
+        const focusedIndex = props.graphState.focusedPointIndex;
+        if (focusedIndex != null) {
+            itemsRef.current[focusedIndex]?.focus();
+        }
+    }, [props.graphState.focusedPointIndex, itemsRef]);
+
     return (
         <>
             {/* This rect is here to grab clicks so that new points can be added */}
@@ -81,11 +90,26 @@ export function UnlimitedPointGraph(props: PointGraphProps) {
                     onMove={(destination) =>
                         dispatch(actions.pointGraph.movePoint(i, destination))
                     }
+                    ref={(ref) => {
+                        itemsRef.current[i] = ref;
+                    }}
                     onFocusChange={(event, isFocused) => {
                         if (isFocused) {
                             dispatch(actions.pointGraph.focusPoint(i));
                         } else {
                             if (event.relatedTarget?.id === REMOVE_BUTTON_ID) {
+                                return;
+                                // This is an optimization: If the next target
+                                // is a point then don't blur because it casues
+                                // the remove button to get taken off the page
+                                // and then put back on The new point will
+                                // receive focus and set the correct state in
+                                // the reducer
+                            } else if (
+                                event.relatedTarget?.classList.contains(
+                                    "movable-point",
+                                )
+                            ) {
                                 return;
                             }
                             dispatch(actions.pointGraph.blurPoint());