+++ packages/perseus/src/widgets/interactive-graphs/reducer/interactive-graph-reducer.ts
@@ -15,7 +15,17 @@ import {
     vector,
 } from "../../../util/geometry";
 import {getQuadraticCoefficients} from "../graphs/quadratic";
-import {clamp, clampToBox, findAngle, inset, polar, snap, X, Y} from "../math";
+import {
+    clamp,
+    clampToBox,
+    getAngleFromVertex,
+    getClockwiseAngle,
+    inset,
+    polar,
+    snap,
+    X,
+    Y,
+} from "../math";
 import {bound} from "../utils";
 
 import {initializeGraphState} from "./initialize-graph-state";
@@ -151,11 +161,16 @@ function doBlurPoint(
 ): InteractiveGraphState {
     switch (state.type) {
         case "point":
-            return {
+            const nextState = {
                 ...state,
-                focusedPointIndex: null,
                 showRemovePointButton: false,
             };
+
+            if (state.interactionMode === "mouse") {
+                nextState.focusedPointIndex = null;
+            }
+
+            return nextState;
         default:
             return state;
     }
@@ -666,6 +681,7 @@ function doAddPoint(
         hasBeenInteractedWith: true,
         coords: [...state.coords, snappedPoint],
         showRemovePointButton: false,
+        focusedPointIndex: state.coords.length,
     };
 }
 
@@ -677,10 +693,18 @@ function doRemovePoint(
         return state;
     }
 
+    let nextFocusedPointIndex: number | null;
+    if (state.interactionMode === "mouse") {
+        nextFocusedPointIndex = null;
+    } else {
+        nextFocusedPointIndex =
+            state.coords.length > 1 ? state.coords.length - 2 : null;
+    }
+
     return {
         ...state,
         coords: state.coords.filter((_, i) => i !== action.index),
-        focusedPointIndex: null,
+        focusedPointIndex: nextFocusedPointIndex,
         showRemovePointButton: false,
     };
 }
@@ -774,7 +798,7 @@ function boundAndSnapAngleVertex(
         const oldPoint = coordsCopy[i];
         let newPoint = vec.add(oldPoint, delta);
 
-        let angle = findAngle(newVertex, newPoint);
+        let angle = getAngleFromVertex(newVertex, newPoint);
         angle *= Math.PI / 180;
 
         newPoint = constrainToBoundsOnAngle(newPoint, angle, range, snapStep);
@@ -892,7 +916,7 @@ function boundAndSnapAngleEndPoints(
     const vertex = coords[1];
 
     // Gets the angle between the coords and the vertex
-    let angle = findAngle(coordsCopy[index], vertex);
+    let angle = getAngleFromVertex(coordsCopy[index], vertex);
 
     // Snap the angle to the nearest multiple of snapDegrees (if provided)
     angle = Math.round((angle - offsetDegrees) / snap) * snap + offsetDegrees;
@@ -954,12 +978,12 @@ function boundAndSnapToPolygonAngle(
     });
 
     const getAngle = function (a: number, vertex, b: number) {
-        const angle = findAngle(
+        const angle = getClockwiseAngle([
             coordsCopy[rel(a)],
-            coordsCopy[rel(b)],
             coordsCopy[rel(vertex)],
-        );
-        return (angle + 360) % 360;
+            coordsCopy[rel(b)],
+        ]);
+        return angle;
     };
 
     const innerAngles = [
@@ -1001,7 +1025,10 @@ function boundAndSnapToPolygonAngle(
         knownSide;
 
     // Angle at the second vertex of the polygon
-    const outerAngle = findAngle(coordsCopy[rel(1)], coordsCopy[rel(-1)]);
+    const outerAngle = getAngleFromVertex(
+        coordsCopy[rel(1)],
+        coordsCopy[rel(-1)],
+    );
 
     // Uses the length of the side of the polygon (radial coordinate)
     // and the angle between the first and second sides of the
@@ -1064,7 +1091,10 @@ function boundAndSnapToSides(
     const innerAngle = lawOfCosines(sides[0], sides[2], sides[1]);
 
     // Angle at the second vertex of the polygon
-    const outerAngle = findAngle(coordsCopy[rel(1)], coordsCopy[rel(-1)]);
+    const outerAngle = getAngleFromVertex(
+        coordsCopy[rel(1)],
+        coordsCopy[rel(-1)],
+    );
 
     // Returns true if the points form a counter-clockwise turn;
     // a.k.a if the point is on the left or right of the polygon.